moving all excluded tests under Core folder
