+++ src/NServiceBus.AcceptanceTests/Core/FakeTransport/FakeDispatcher.cs
@@ -1,4 +1,4 @@
-namespace NServiceBus.AcceptanceTests.FakeTransport
+namespace NServiceBus.AcceptanceTests.Core.FakeTransport
 {
     using System.Threading.Tasks;
     using Extensibility;
