+++ src/NServiceBus.AcceptanceTests/Core/FakeTransport/FakeTransport.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.FakeTransport
+﻿namespace NServiceBus.AcceptanceTests.Core.FakeTransport
 {
     using Settings;
     using Transport;
