+++ src/NServiceBus.AcceptanceTests/Core/FakeTransport/FakeTransportInfrastructure.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.FakeTransport
+﻿namespace NServiceBus.AcceptanceTests.Core.FakeTransport
 {
     using System;
     using System.Collections.Generic;
