+++ src/NServiceBus.AcceptanceTests/Core/Persistence/When_a_persistence_does_not_support_saga.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.Persistence
+﻿namespace NServiceBus.AcceptanceTests.Core.Persistence
 {
     using System;
     using System.Threading.Tasks;
