+++ src/NServiceBus.AcceptanceTests/Core/Persistence/When_a_persistence_does_not_support_timeouts.cs
@@ -1,12 +1,14 @@
-namespace NServiceBus.AcceptanceTests.Persistence
+﻿namespace NServiceBus.AcceptanceTests.Core.Persistence
 {
+    using System.Threading.Tasks;
     using AcceptanceTesting;
     using EndpointTemplates;
+    using Features;
     using NServiceBus.Persistence;
     using NUnit.Framework;
     using ScenarioDescriptors;
 
-    public class When_a_persistence_does_not_support_subscriptions : NServiceBusAcceptanceTest
+    public class When_a_persistence_does_not_support_timeouts : NServiceBusAcceptanceTest
     {
         [Test]
         public void should_throw_exception()
@@ -14,10 +16,10 @@ namespace NServiceBus.AcceptanceTests.Persistence
             Assert.That(async () =>
             {
                 await Scenario.Define<Context>()
-                    .WithEndpoint<Endpoint>(e => e.When(b => b.Subscribe<object>()))
-                    .Repeat(r => r.For<AllTransportsWithMessageDrivenPubSub>())
+                    .WithEndpoint<Endpoint>(e => e.When(b => Task.FromResult(0)))
+                    .Repeat(r => r.For<AllTransportsWithoutNativeDeferral>())
                     .Run();
-            }, Throws.Exception.InnerException.InnerException.With.Message.Contains("DisableFeature<MessageDrivenSubscriptions>()"));
+            }, Throws.Exception.InnerException.InnerException.With.Message.Contains("DisableFeature<TimeoutManager>()"));
         }
 
         class Endpoint : EndpointConfigurationBuilder
@@ -29,7 +31,9 @@ namespace NServiceBus.AcceptanceTests.Persistence
                     c.UsePersistence<InMemoryPersistence, StorageType.Sagas>();
                     c.UsePersistence<InMemoryPersistence, StorageType.GatewayDeduplication>();
                     c.UsePersistence<InMemoryPersistence, StorageType.Outbox>();
-                    c.UsePersistence<InMemoryPersistence, StorageType.Timeouts>();
+                    c.UsePersistence<InMemoryPersistence, StorageType.Subscriptions>();
+
+                    c.EnableFeature<TimeoutManager>();
                 });
             }
         }
