+++ src/NServiceBus.AcceptanceTests/Core/UnitOfWork/TransactionScope/When_using_timeout_greater_than_machine_max.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.UnitOfWork
+﻿namespace NServiceBus.AcceptanceTests.Core.UnitOfWork.TransactionScope
 {
     using System;
     using AcceptanceTesting;
