+++ src/NServiceBus.AcceptanceTests/Core/UnitOfWork/When_a_subscription_message_arrives.cs
@@ -1,4 +1,4 @@
-﻿namespace NServiceBus.AcceptanceTests.UnitOfWork
+﻿namespace NServiceBus.AcceptanceTests.Core.UnitOfWork
 {
     using System;
     using System.Threading.Tasks;
