+++ src/NServiceBus.AcceptanceTests/DelayedDelivery/TimeoutManager/When_timeout_storage_fails.cs
@@ -7,8 +7,9 @@
     using EndpointTemplates;
     using Extensibility;
     using Features;
-    using NServiceBus.Persistence;
+    using NServiceBus;
     using NUnit.Framework;
+    using Persistence;
     using ScenarioDescriptors;
     using Timeout.Core;
