+++ src/NServiceBus.AcceptanceTests/NServiceBus.AcceptanceTests.csproj
@@ -73,14 +73,14 @@
     <Compile Include="DelayedDelivery\TimeoutManager\When_timeout_storage_fails.cs" />
     <Compile Include="EndpointTemplates\EndpointCustomizationConfigurationExtensions.cs" />
     <Compile Include="EndpointTemplates\ServerWithNoDefaultPersistenceDefinitions.cs" />
-    <Compile Include="FakeTransport\CriticalError\When_registering_custom_critical_error_handler.cs" />
-    <Compile Include="FakeTransport\FakeDispatcher.cs" />
-    <Compile Include="FakeTransport\FakeQueueCreator.cs" />
-    <Compile Include="FakeTransport\FakeReceiver.cs" />
-    <Compile Include="FakeTransport\FakeTransport.cs" />
-    <Compile Include="FakeTransport\FakeTransportInfrastructure.cs" />
-    <Compile Include="FakeTransport\FakeTransportSetingsExtensions.cs" />
-    <Compile Include="FakeTransport\ProcessingOptimizations\When_using_concurrency_limit.cs" />
+    <Compile Include="Core\FakeTransport\CriticalError\When_registering_custom_critical_error_handler.cs" />
+    <Compile Include="Core\FakeTransport\FakeDispatcher.cs" />
+    <Compile Include="Core\FakeTransport\FakeQueueCreator.cs" />
+    <Compile Include="Core\FakeTransport\FakeReceiver.cs" />
+    <Compile Include="Core\FakeTransport\FakeTransport.cs" />
+    <Compile Include="Core\FakeTransport\FakeTransportInfrastructure.cs" />
+    <Compile Include="Core\FakeTransport\FakeTransportSetingsExtensions.cs" />
+    <Compile Include="Core\FakeTransport\ProcessingOptimizations\When_using_concurrency_limit.cs" />
     <Compile Include="Basic\When_depending_on_typed_feature.cs" />
     <Compile Include="Hosting\When_overriding_local_addresses.cs" />
     <Compile Include="Hosting\When_a_message_is_faulted.cs" />
@@ -102,13 +102,13 @@
     <Compile Include="Core\MessageDurability\When_disabling_durable_messages_on_durable_transport.cs" />
     <Compile Include="Core\MessageDurability\When_using_non_durable_messages_on_durable_only_transport.cs" />
     <Compile Include="Performance\When_message_is_faulted.cs" />
-    <Compile Include="Persistence\When_a_persistence_does_not_support_outbox.cs" />
-    <Compile Include="Persistence\When_a_persistence_does_not_support_saga.cs" />
-    <Compile Include="Persistence\When_a_persistence_does_not_support_subscriptions.cs" />
-    <Compile Include="Persistence\When_a_persistence_does_not_support_timeouts.cs" />
+    <Compile Include="Core\Persistence\When_a_persistence_does_not_support_outbox.cs" />
+    <Compile Include="Core\Persistence\When_a_persistence_does_not_support_saga.cs" />
+    <Compile Include="Core\Persistence\When_a_persistence_does_not_support_subscriptions.cs" />
+    <Compile Include="Core\Persistence\When_a_persistence_does_not_support_timeouts.cs" />
     <Compile Include="Pipeline\When_using_per_uow_component_in_the_pipeline.cs" />
     <Compile Include="Pipeline\When_replacing_behavior.cs" />
-    <Compile Include="Persistence\When_a_persistence_does_not_provide_ISynchronizationContext.cs" />
+    <Compile Include="Core\Persistence\When_a_persistence_does_not_provide_ISynchronizationContext.cs" />
     <Compile Include="Recoverability\Retries\When_message_is_deferred_by_slr_using_dtc.cs" />
     <Compile Include="Recoverability\Retries\When_message_fails_retries.cs" />
     <Compile Include="Recoverability\Retries\When_performing_slr_with_flr_disabled.cs" />
@@ -216,11 +216,11 @@
     <Compile Include="Tx\ImmediateDispatch\When_requesting_immediate_dispatch_with_exactly_once.cs" />
     <Compile Include="Tx\Issue_2481.cs" />
     <Compile Include="Tx\When_receiving_with_native_multi_queue_transaction.cs" />
-    <Compile Include="UnitOfWork\TransactionScope\When_transactionscope_enabled.cs" />
-    <Compile Include="UnitOfWork\TransactionScope\When_used_with_default_transaction_mode.cs" />
-    <Compile Include="UnitOfWork\TransactionScope\When_using_timeout_greater_than_machine_max.cs" />
-    <Compile Include="UnitOfWork\TransactionScope\When_used_with_transport_scopes.cs" />
-    <Compile Include="UnitOfWork\When_a_subscription_message_arrives.cs" />
+    <Compile Include="Core\UnitOfWork\TransactionScope\When_transactionscope_enabled.cs" />
+    <Compile Include="Core\UnitOfWork\TransactionScope\When_used_with_default_transaction_mode.cs" />
+    <Compile Include="Core\UnitOfWork\TransactionScope\When_using_timeout_greater_than_machine_max.cs" />
+    <Compile Include="Core\UnitOfWork\TransactionScope\When_used_with_transport_scopes.cs" />
+    <Compile Include="Core\UnitOfWork\When_a_subscription_message_arrives.cs" />
     <Compile Include="Volatile\When_sending_to_non_durable_endpoint.cs" />
     <Compile Include="Encryption\When_using_Rijndael_with_multikey.cs" />
     <Compile Include="Encryption\When_using_Rijndael_with_custom.cs" />
